requirements.txt
flask
gunicorn
requests
pymongo[srv]
python-dotenv
html
redis
threading
json